Get Tomic Table

Extract one of the specific tables from a tomic object

get_tomic_table(tomic, table_type)

Arguments

tomic

Either a tidy_omic or triple_omic object

table_type

The type of table to extract from the tomic object.

tidy

one row per measurements with feature and sample attributes added. Equivalent to the $data field of a tidy omic object

measurements

one row per measurements defined a feature and sample foreign key. Equivalent to the $measurements field of a triple omic object

features

one row per feature defined by a feature primary key. Equivalent to the $features field of a triple omic object

samples

one row per sample defined by a sample primary key. Equivalent to the $samples field of a triple omic object

Value

a tibble matching the table_type of the tomic object

Examples

get_tomic_table(brauer_2008_triple, "samples")
#> # A tibble: 36 × 3
#>    sample nutrient    DR
#>    <chr>  <chr>    <dbl>
#>  1 G0.05  G         0.05
#>  2 G0.1   G         0.1 
#>  3 G0.15  G         0.15
#>  4 G0.2   G         0.2 
#>  5 G0.25  G         0.25
#>  6 G0.3   G         0.3 
#>  7 N0.05  N         0.05
#>  8 N0.1   N         0.1 
#>  9 N0.15  N         0.15
#> 10 N0.2   N         0.2 
#> # … with 26 more rows
get_tomic_table(brauer_2008_tidy, "features")
#> # A tibble: 500 × 4
#>    name    BP                                                      MF    syste…¹
#>    <chr>   <chr>                                                   <chr> <chr>  
#>  1 YOL029C biological process unknown                              mole… YOL029C
#>  2 SCW11   cytokinesis, completion of separation                   gluc… YGL028C
#>  3 YHR036W biological process unknown                              mole… YHR036W
#>  4 BGL2    cell wall organization and biogenesis                   gluc… YGR282C
#>  5 ACT1    cell wall organization and biogenesis*                  stru… YFL039C
#>  6 FKH1    pseudohyphal growth*                                    tran… YIL131C
#>  7 HOC1    cell wall mannoprotein biosynthesis*                    tran… YJR075W
#>  8 CSN12   adaptation to pheromone during conjugation with cellul… mole… YJR084W
#>  9 YAL046C biological process unknown                              mole… YAL046C
#> 10 SLG1    cell wall organization and biogenesis*                  tran… YOR008C
#> # … with 490 more rows, and abbreviated variable name ¹​systematic_name