Remove Missing Values

Account for missing values by dropping features, samples or using imputation.

remove_missing_values(
  tomic,
  value_var = NULL,
  missing_val_method = "drop_samples"
)

Arguments

tomic

Either a tidy_omic or triple_omic object

value_var

An abundance value to use with hclust

missing_val_method

Approach to remove missing values:

drop_features

Drop features with missing values

drop_samples

Drop samples which are missing all features, then drop features

impute

Impute missing values

Value

A tomic object where missing values have been accounted for.

Examples

remove_missing_values(brauer_2008_triple)
#> 40 features dropped due to missing values
#> $features
#> # A tibble: 460 × 4
#>    name    BP                                                      MF    syste…¹
#>    <chr>   <chr>                                                   <chr> <chr>  
#>  1 YOL029C biological process unknown                              mole… YOL029C
#>  2 SCW11   cytokinesis, completion of separation                   gluc… YGL028C
#>  3 YHR036W biological process unknown                              mole… YHR036W
#>  4 BGL2    cell wall organization and biogenesis                   gluc… YGR282C
#>  5 ACT1    cell wall organization and biogenesis*                  stru… YFL039C
#>  6 FKH1    pseudohyphal growth*                                    tran… YIL131C
#>  7 HOC1    cell wall mannoprotein biosynthesis*                    tran… YJR075W
#>  8 CSN12   adaptation to pheromone during conjugation with cellul… mole… YJR084W
#>  9 YAL046C biological process unknown                              mole… YAL046C
#> 10 SLG1    cell wall organization and biogenesis*                  tran… YOR008C
#> # … with 450 more rows, and abbreviated variable name ¹​systematic_name
#> 
#> $samples
#> # A tibble: 36 × 3
#>    sample nutrient    DR
#>    <chr>  <chr>    <dbl>
#>  1 G0.05  G         0.05
#>  2 G0.1   G         0.1 
#>  3 G0.15  G         0.15
#>  4 G0.2   G         0.2 
#>  5 G0.25  G         0.25
#>  6 G0.3   G         0.3 
#>  7 N0.05  N         0.05
#>  8 N0.1   N         0.1 
#>  9 N0.15  N         0.15
#> 10 N0.2   N         0.2 
#> # … with 26 more rows
#> 
#> $measurements
#> # A tibble: 16,560 × 3
#>    name    sample expression
#>    <chr>   <chr>       <dbl>
#>  1 YOL029C G0.05       -0.22
#>  2 SCW11   G0.05       -0.67
#>  3 YHR036W G0.05       -0.91
#>  4 BGL2    G0.05       -0.08
#>  5 ACT1    G0.05       -0.04
#>  6 FKH1    G0.05       -0.57
#>  7 HOC1    G0.05       -0.42
#>  8 CSN12   G0.05       -0.49
#>  9 YAL046C G0.05        0.05
#> 10 SLG1    G0.05       -0.06
#> # … with 16,550 more rows
#> 
#> $design
#> $design$features
#> # A tibble: 4 × 2
#>   variable        type               
#>   <chr>           <chr>              
#> 1 name            feature_primary_key
#> 2 systematic_name character          
#> 3 BP              character          
#> 4 MF              character          
#> 
#> $design$samples
#> # A tibble: 3 × 2
#>   variable type              
#>   <chr>    <chr>             
#> 1 sample   sample_primary_key
#> 2 nutrient character         
#> 3 DR       numeric           
#> 
#> $design$measurements
#> # A tibble: 3 × 2
#>   variable   type               
#>   <chr>      <chr>              
#> 1 name       feature_primary_key
#> 2 sample     sample_primary_key 
#> 3 expression numeric            
#> 
#> $design$feature_pk
#> [1] "name"
#> 
#> $design$sample_pk
#> [1] "sample"
#> 
#> 
#> attr(,"class")
#> [1] "triple_omic" "tomic"       "general"